+++ Terminaltor/Terminaltor/GameObject.cpp
@@ -9,7 +9,7 @@ m_sprite()
 {
 }
 
-GameObject::GameObject( int x, int y, int width, int height ) :
+GameObject::GameObject( float x, float y, int width, int height ) :
 m_x( x ),
 m_y( y ),
 m_width( width ),
