+++ Terminaltor/Terminaltor/GameState.h
@@ -13,8 +13,8 @@ public:
 	static inline void SetState(int state) { m_state = state; };
 	static inline int Fps() { return m_fps; };
 	static inline void SetFps(int fps) { m_fps = fps; };
-	static inline int Delay() { return m_delay; };
-	static inline void SetDelay(int delay) { m_delay = delay; };
+	static inline float Delay() { return m_delay; };
+	static inline void SetDelay(float delay) { m_delay = delay; };
 	static inline Menu* MainMenu() { return m_mainMenu; };
 	static inline void SetMainMenu( Menu* menu ) { m_mainMenu = menu; };
 
@@ -25,6 +25,6 @@ private:
 	static int m_state;
 	static int m_fps;
 	static Menu* m_mainMenu;
-	static int m_delay;
+	static float m_delay;
 };
